+++ R/globals.R
@@ -6,3 +6,6 @@ utils::globalVariables(c('pip', 'id', 'label', 'WAIC', 'zscore', 'geom_text_repe
 utils::globalVariables(c('prior_p', 'cluster_id', 'group'))
 ## chain_seed is supplied to each worker via future(..., globals =) in ivd()
 utils::globalVariables('chain_seed')
+## globals for pp_check.ivd (ggplot aes columns)
+utils::globalVariables(c('ordered', 'lo_outer', 'hi_outer', 'lo_inner',
+                         'hi_inner', 'med', 'obs', 'outside', 'value', 'rep_id'))